Split SIMD into separate files
